Update unit tests for DidResolver.spec.ts
